requirements.txt
streamlit==1.37.1
openai==1.42.0
DateTime==5.5
pandas==2.2.2
python-dotenv==1.0.0